chore: move to controller/hybridgateway/metadata/annotation.go
